# docs/swagger.yaml
basePath: /api/v3
definitions:
  pricing.Config:
    properties:
      base_prices:
        $ref: '#/definitions/pricing.PriceByTypeUSD'
      country_modifiers:
        additionalProperties:
          $ref: '#/definitions/pricing.Modifier'
        type: object
    type: object
  pricing.LatestPrices:
    properties:
      current_valid_until:
        type: string
      defaults:
        $ref: '#/definitions/pricing.PriceHistory'
      per_country:
        additionalProperties:
          $ref: '#/definitions/pricing.PriceHistory'
        type: object
      previous_valid_until:
        type: string
    type: object
  pricing.Modifier:
    properties:
      other:
        type: number
      residential:
        type: number
    type: object
  pricing.Price:
    properties:
      price_per_gib:
        type: integer
      price_per_hour:
        type: integer
    type: object
  pricing.PriceByType:
    properties:
      other:
        $ref: '#/definitions/pricing.Price'
      residential:
        $ref: '#/definitions/pricing.Price'
    type: object
  pricing.PriceByTypeUSD:
    properties:
      other:
        $ref: '#/definitions/pricing.PriceUSD'
      residential:
        $ref: '#/definitions/pricing.PriceUSD'
    type: object
  pricing.PriceHistory:
    properties:
      current:
        $ref: '#/definitions/pricing.PriceByType'
      previous:
        $ref: '#/definitions/pricing.PriceByType'
    type: object
  pricing.PriceUSD:
    properties:
      price_per_gib_usd:
        type: number
      price_per_hour_usd:
        type: number
    type: object
  proposal.PingResponse:
    properties:
      message:
        type: string
    type: object
  v2.AccessPolicy:
    properties:
      id:
        type: string
      source:
        type: string
    type: object
  v2.Contact:
    properties:
      definition:
        type: object
      type:
        type: string
    type: object
  v2.Location:
    properties:
      asn:
        type: integer
      city:
        type: string
      continent:
        type: string
      country:
        type: string
      ip_type:
        type: string
      isp:
        type: string
    type: object
  v2.Metadata:
    properties:
      country:
        type: string
      ip_type:
        type: string
      isp:
        type: string
      price_per_gib:
        type: integer
      price_per_hour:
        type: integer
      provider_id:
        type: string
      service_type:
        type: string
      updated_at:
        type: string
      whitelist:
        type: boolean
    type: object
  v2.Price:
    properties:
      currency:
        type: string
      per_gib:
        type: integer
      per_hour:
        type: integer
    type: object
  v2.Proposal:
    properties:
      access_policies:
        items:
          $ref: '#/definitions/v2.AccessPolicy'
        type: array
      compatibility:
        type: integer
      contacts:
        items:
          $ref: '#/definitions/v2.Contact'
        type: array
      format:
        type: string
      id:
        type: integer
      location:
        $ref: '#/definitions/v2.Location'
      price:
        $ref: '#/definitions/v2.Price'
      provider_id:
        type: string
      quality:
        $ref: '#/definitions/v2.Quality'
      service_type:
        type: string
    type: object
  v2.Quality:
    properties:
      bandwidth:
        description: Bandwidth in Mbps.
        type: number
      latency:
        description: Latency in ms.
        type: number
      quality:
        description: Quality valuation from the oracle.
        type: number
    type: object
info:
  contact: {}
  description: Discovery API for Mysterium Network
  title: Discovery API
  version: "3.0"
paths:
  /ping:
    get:
      consumes:
      - application/json
      description: Ping
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/proposal.PingResponse'
      summary: Ping
      tags:
      - system
  /prices:
    get:
      description: Latest Prices
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/pricing.LatestPrices'
            type: array
      summary: Latest Prices
      tags:
      - prices
  /prices/config:
    get:
      description: price config
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/pricing.Config'
            type: array
      summary: Price config
      tags:
      - prices
    post:
      description: update price config
      parameters:
      - description: config object
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/pricing.Config'
      responses:
        "202":
          description: ""
      summary: update price config
      tags:
      - prices
  /proposals:
    get:
      consumes:
      - application/json
      description: List proposals
      parameters:
      - description: Consumer country
        in: query
        name: from
        type: string
      - description: Provider ID
        in: query
        name: provider_id
        type: string
      - description: Service type
        in: query
        name: service_type
        type: string
      - description: Provider country
        in: query
        name: location_country
        type: string
      - description: IP type (residential, datacenter, etc.)
        in: query
        name: ip_type
        type: string
      - description: Access policy. When empty, returns only public proposals (default).
          Use 'all' to return all.
        in: query
        name: access_policy
        type: string
      - description: Access policy source
        in: query
        name: access_policy_source
        type: string
      - description: Maximum price per GiB. When empty, will not filter by it. Price
          is set in ethereum wei.
        in: query
        name: price_gib_max
        type: number
      - description: Maximum price per hour. When empty, will not filter by it. Price
          is set in ethereum wei.
        in: query
        name: price_hour_max
        type: number
      - description: Minimum compatibility. When empty, will not filter by it.
        in: query
        name: compatibility_min
        type: number
      - description: Maximum compatibility. When empty, will not filter by it.
        in: query
        name: compatibility_max
        type: number
      - description: Minimal quality threshold. When empty will be defaulted to 0.
          Quality ranges from [0.0; 3.0]
        in: query
        name: quality_min
        type: number
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v2.Proposal'
            type: array
      summary: List proposals
      tags:
      - proposals
  /proposals-metadata:
    get:
      consumes:
      - application/json
      description: List proposals' metadata
      parameters:
      - description: Provider ID
        in: query
        name: provider_id
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v2.Metadata'
            type: array
      summary: List proposals' metadata.
swagger: "2.0"